chore: update accordion
