+++ angular.json
@@ -72,13 +72,6 @@
               "browserTarget": "ngx-weui-doc:build:production"
             }
           }
-        },
-        "lint": {
-          "builder": "@angular-devkit/build-angular:tslint",
-          "options": {
-            "tsConfig": ["src/tsconfig.json"],
-            "exclude": ["**/node_modules/**"]
-          }
         }
       }
     },
