+++ components/accordion/accordion.module.ts
@@ -1,8 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 
-import { AccordionPanelComponent } from './accordion-panel.component';
-import { AccordionComponent } from './accordion.component';
+import { AccordionComponent, AccordionPanelComponent } from './accordion.component';
 
 @NgModule({
   imports: [CommonModule],
